backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.22.0
motor==3.7.0
beanie[odm]==2.0.0
pydantic==2.11.7
email-validator>=2.0
python-jose[cryptography]==3.3.0
python-dotenv==1.1.1
bcrypt==4.0.1
passlib==1.7.4